Move unlinking directly into the purge action
